Keep raftId as a uint16 to support compatibility.
In order to support an easy upgrade with no downtime, e.g.  node by node,
and to be able to add new nodes to an existing network, having quorum remain
backwards compatible with existing network, the raftId needs to  remain an uint16
(on the quorum side) instead of an uint64 for now. It is always a
uint64 on the raft etcd protocol level.

note: Making the raftId a uint64 is
tricky because when an old node sends a raft peer address over the wire,
it has a type of uint16, so in order to change the type to an uint16 we
would have to support deserializing an uint16 and then if that failed an
uint64. All nodes would have to be upgraded with the code to catch both
cases.

+++ raft/api.go
@@ -123,14 +123,14 @@ func (s *PublicRaftAPI) Cluster() ([]ClusterInfo, error) {
 		if !noLeader {
 			if a.RaftId == leaderAddr.RaftId {
 				role = "minter"
-			} else if s.raftService.raftProtocolManager.isLearner(a.RaftId) {
+			} else if s.raftService.raftProtocolManager.isLearner(uint64(a.RaftId)) {
 				role = "learner"
-			} else if s.raftService.raftProtocolManager.isVerifier(a.RaftId) {
+			} else if s.raftService.raftProtocolManager.isVerifier(uint64(a.RaftId)) {
 				role = "verifier"
 			}
 		}
 		//log.Info("Cluster RaftId is: ", "a.RaftId", a.RaftId, "aJS.RaftId", aJS.RaftId)
-		clustInfo[i] = ClusterInfo{*aJS, role, s.checkIfNodeIsActive(a.RaftId)}
+		clustInfo[i] = ClusterInfo{*aJS, role, s.checkIfNodeIsActive(uint64(a.RaftId))}
 	}
 	return clustInfo, nil
 }
